Move editor rendering to a separate view
